Lexer and parser fixes.
Fixing number parsing in the lexer.  Currently the lexer will parse
"123foo" as two tokens: the number 123 and a bare word "foo". I've tried
specifying the number must be on a word boundary, but that seems to
prevent it from matching a number entirely.  As a workaround, adding a
token to the lexer which means "invalid number", and we match this token
for any number that ends in any number of non-digit characters.

Adding a proper error handler type that is responsible for logging
warnings and errors.

Extracting the logic in main into a "parser".

Moving some definitions out of lexer.hpp so that AST files don't have to
include all of lexer just for defining token positions.

Fixed invalid escape sequences in strings being handled as an error.
They are now properly handled as warnings.

Implemented missing functionality to parse from manifest strings.