Implement function calls.
Implementing function calls and yielding to lambdas.

The following Puppet functions are now implemented:

* alert
* crit
* debug
* emerg
* err
* fail
* info
* notice
* warning
* with

Lambda parameter type validation is not still unimplemented.

This commit also changes the output format to better match Puppet's.

Additionally, the AST is now "read-only".  This was needed to allow
lambda bodies to be executed more than once.