Fix interpolation of variable names with leading underscores.
Fixing string interpolation that should handle the conversion of bare
words (may lead with underscore) into variables.  This fixes
interpolation of ${_foo}, ${_foo[0]} and ${_foo.bar}, where _foo should
be treated as a reference to variable $_foo.

This also fixes the parsing of bare words, which caused an unhandled
exception since string interpolation was implemented.  It also changes
the grammar such that the parser will now output "bare words" into the
AST instead of treating them as strings.  This allows us to handle the
above expressions.

This addresses GH issue #9.