Prevent unproductive expression evaluation.
This commit makes the expression evaluator behave more like Puppet's
with respect to evaluating unproductive expressions.

For example, '1+1' is an unproductive expression because it doesn't
change the state of the system.  Puppet detects these expressions and
treats them like an error, provided the expression isn't the last in a
block (which functions like a "return value").

This checking is actually more strict than Puppet's.  For example, the
expression '1 + ($a = 1)' is treated as productive by Puppet because the
subexpression $a = 1 is productive.  This implementation considers that
to be an error since the entire expression is unproductive even if the
subexpression is productive.