Implement method call expression evaluation.
Implementing method call expression evaluation.  Functions can now be
called with the "method call" syntax that passes the target as the first
argument to the function:

with(*[1, 2, 3]) |$x, $y, $z| {
    notice $x, $y, $z
}

is the same as:

*[1, 2, 3].with |$x, $y, $z| {
    notice $x, $y, $z
}

Note: Puppet 4's compiler does not properly unfold the above in the
method call syntax.

This also fixes splat of nested expressions, e.g. `notice((*[1, 2,
3]))`.