Treat 'default' like 'undef'.
Currently default is special cased in certain expressions and cannot be
used outside of those expressions.  Puppet treats the default keyword as
part of the type system (has type 'Default'), much like undef.  This
commit changes the grammar, AST, and expression evaluator to add a
"defaulted" (named such due to "default" being a reserved word in C++)
value so that it can participate in the type system and appear in any
expression.