Fix postfix expressions.
Previously the grammar was defined such that postfix expressions could
only chain with expressions of the same type.

For example, access expressions could only chain together with other
access expressions.  So $a[0][0] was a valid expression, but "$a.with
|$x| { $x }[0]" was not.

This fixes the grammar such that postfix expressions are primary
expressions that have an initial primary expression and then N number of
postfix expressions that are evaluated in sequence, chained together.

This also fixes a bug in lambdas such that lambdas returning a match or
parameter variable (i.e. a variable in an ephemeral scope) failed to
copy the value into the return value.  This resulted in undefined
behavior when the variable was later accessed.