Refactor operators into seperate files.
Refactoring the operators into their own files.

Fixing usages of `boost::get` on a value; this should only be valid on a
value when checking to see if it is a variable.  Otherwise, we should
use `as<T>` to check for the value's type and use as-is.  To make changes to a
value, we should use `mutate` and `mutate_as`.  These functions will
make necessary copies if the value is a variable; if not a variable,
then the original value is returned as-is (for `mutate_as`, the value
must be the same type as T).

This will fix places where variables cause unexpected behavior
because they weren't being dereferenced.