Fix case expression parsing.
The previous commit moved resource expressions out of statement
expressions and into primary expressions.  This caused ambiguity in
case expressions.

For example, this could be parsed as a case keyword followed by a
resource expression for a resource of type 'foo':

```
case foo {
} ...
```

The only catalog expression that should be allowed a primary expression
is the collection expression; all other catalog expressions are treated
as statement expressions.