Implement "safer" rvalue_cast.
One of the problems with `std::move` is that it will case a `const&` to
`const&&`, which in turn will bind to copy constructors and copy-assignment
operators over move constructors and move-assignment operators.

This implements `puppet::rvalue_cast` that performs the same function as
`std::move` except that it asserts that the argument that was provided
is not const.  If it is const, it causes a compilation error.

This should hopefully catch places where a move was intended but a copy
was actually performed because the argument was const.