Refactor AST to allow for eventual move semantics.
Originally the AST was written to be read-only. However, for evaluating
expressions such as class and defined type, there needs to be a way to
"save" part of the AST for evaluation later.  An efficient way to do
that would be to move those expressions into a container for later
evaluation.

This changes a few things:

* All AST types have removed the read-only interface in favor of a
  POD-like interface of just storing the data, allowing move semantics
  of contained data.
* puppet::lexer::token_position has simply renamed to
  puppet::lexer::position.
* The grammar has been cleaned up to remove unnecessary Phoenix
  constructions.

This change does not yet update the expression evaluator to allow for
move semantics of parts of the AST; that will come with the
implementation of the aforementioned catalog expressions.