Revert previous refactoring of AST types.
Previously a refactoring was done to the AST types in an attempt to make
the types movable.  The intention was that certain catalog expressions
would move parts of the AST into lists to make use of later (such as
realizing classes and defined types).

However, this prevents efficient repeat execution of expressions because
they must be mutable.  For functions and lambdas, this means that every
expression must be copied before being executed so that the AST remains
unmodified.

Rather than mutating the AST, the forthcoming design is to have the
parser return a shared pointer to the AST.  When part of an AST needs to
be used later, the shared pointer will be stored along with an interior
pointer to the relevant part of the AST.