Implement expression short circuiting for logical operators.
Implementing evaluation short circuiting for logical and/or operators.

If the operator is logical `and` and the LHS evaluates to false, then the
RHS is not evaluated at all. The result of the expression is false.

If the operator is logical `or` and the LHS evaluates to true, then the
RHS is not evaluated at all. The result of the expression is true.

This corrects the previous behavior where the RHS would always be
evaluted.  This prevents expressions such as:

```
if $foo and $foo > 5 {
    notice(yes)
}
```

If $foo was undef, the RHS was being evaluated and undef is not a valid
operand to operator `>`, so an error was raised.

With the short circuiting, `$foo > 5` is no longer evaluated when $foo
is undef.