Support Type and Resource['type'] in resource expressions.
Resource expressions in Puppet allow for specifying the type with a
resource Type or a Resource['type'] expression.  Previously the grammar
was defined to only allow a name or the keyword 'class'.  The grammar
has been updated to accept a type or a qualified Resource expression.

Examples:

```
File {
    '/foo':
}

Resource['file'] {
    '/foo':
}

File {
    '/foo':
}

$a = 'file'
Resource[$a] {
    '/foo':
}
```