Fix incorrect invocations of as<T>.
Due to C++ implicitly converting pointer types to bool, invocations of
as<T>(&value) resulted in creating a temporary value set to "true".
This is not the intention of these invocations.

Adding an overload of as<T> that takes a `value const*`.  This overload
is not defined, so any attempt to call as<T> this way will result in an
expected link error.

Fixing a few places where as<T> was being called incorrectly.