Fix class and defined type redeclarations.
Class and defined type redeclarations should be an error.

For classes, it was silently succeeding.  For defined types, the error
message was unhelpful.

This was a regression from a recent refactoring of the catalog
expression evaluator.

This also fixes a place where a pointer was output in an error message
instead of the path it points to.