Various improvements.
Implementing adding Stage[main], Class[main], and Class[settings] to the
catalog.  Settings scope is not yet populated with any values, but is
now present.

Removing complex error handling from executor in favor of exceptions.

Adding more validation of classes and defined types.

Adding containing resource to scope.  This will be used in an upcoming
change to add containment edges to the dependency graph.