Fix defined type declaration not finding types.
The previous commit regressed defined type declarations by accidentally
searching for the defined type by title rather than by type name.  This fixes
that by taking the type name (expected to be lowercase for the map) as a
parameter when declaring the defined type.

Also removing unused catalog member from resource.  If we need a back reference
from resource to catalog, we can add it back in the future.