Fix unhandled exception when puppetcpp is run without arguments.
When run without arguments, the check for argc > 1 fails so the program options
map is not populated correctly with default values.  As a result, casting a
non-existent value to a logging level throws.

The fix is to not check for argc > 1 and modify the default constructor for
`puppet::compiler::settings` to populate the map by parsing a trivial
command line of just "puppetcpp".

This fixes GH-58.