Fix resources with empty bodies not inheriting default attributes.
The following code should print 'foo':

```
file {
  default:
  owner => foo
  ;
  '/foo':
  ;
}

notice File['/foo'][owner]
```

However, it results in an error saying that owner is not a parameter on
File['/foo'] (this itself is a bug, but it cannot be fixed until custom types
are supported).

This fixes the parent pointer of the attributes collection being set when no
attributes are present in a resource body.