Fix operator + with two hashes containing the same key.
Operator + with two hash operands should map the value on the right-hand side
when the key exists in both hashes.  Because insert was being called, existing
keys on the left were retaining the left-hand side values.

The fix is to simply use operator[] to overwrite any keys that already exist.