Fix defined type evaluation.
Previously, defined type evaluation was occurring immediately upon declaration.
However, defined types support the resource override and resource default
expressions.  This means that defined types must be evaluated after everything
else has been evaluated.  In the Ruby compiler, this happens during catalog
finalization.

This fix refactors the defined type support so that it is possible to delay
their evaluations until catalog finalization.  Attributes now store their
contexts and positions so that the compiler can point to either the name or
value of the attribute if there is an error.

This also refactors the "contain", "include", and "require" functions into a
single implementation called "declare", which declares a class and optionally
forms a relationship for the contain/require functions.

Also implemented is the ability to "splat" resource attributes from a hash
using the special "*" attribute name.