Add pointer overloads of value functions.
Adding pointer overloads of functions that should take values by const&.

This is to prevent unintentional calls by pointers to values that are
implicitly convertible to `value const&`.

Because only the definitions are provided, it is expected to break the build
with a linker error if these overloads are invoked.