Fix heredoc positions.
Previously, heredoc positions were the line following the opening heredoc (i.e.
where the string part of the heredoc starts).  However, this can be confusing
as the start of the next line might not contain contents of the string if
whitespace is being trimmed.

This changes the position of the heredoc token to where the heredoc tag is,
i.e. at the start of the @.