Fix leading tabs in source line causing incorrect "error pointer".
When printing the "error pointer" for context into where on a line a warning or
error occurred, any leading tab characters were not taken into account when
calculating how many spaces to write before the pointer.  Therefore the pointer
was offset compared to the source line.

If a leading tab character is encountered, it is now properly converted to a
"tab width" (4) of spaces when printing the "pointer".