Implement support for nested expressions in collector queries.
Previously the grammar didn't support having nested query expressions in
collector queries.  Therefore, this would not compile:

```
File <| tag == foo and (owner == bar or owner == baz) |>
```

This would only work without the parenthesis, which does not have the same
semantic meaning.

This modifies the AST and grammar to allow for attribute queries to be
nested as expected.