Implementing lazy evaluation of resource overrides and relationships.
In Puppet, resource override and relationship expressions are order
independent.  That means they should not be immediately evaluated if the
resource does not yet exist in the catalog.

For resource override expressions, this means storing the context of the
resource override for later evaluation if it does not yet exist.  The overrides
are evaluated when the resource is declared or at during catalog finalization.

For resource relationships, the relationship is stored in a list and processed
in order when the catalog is finalized.  This also changes the resources to
populate the relationship metaparameters based on the state of the dependency
graph rather than merging array values.

A few other minor fixes are contained in this commit.