Fix caching of whether or not module manifests have loaded.
When loading a manifest for 'foo::bar::baz', the resolved path to the manifest
was being cached instead of the name that was passed to `load_manifest`.  This
resulted in needlessly resolving the path to the manifest file when it had
already been searched for previously.

The fix was to cache the given name before the resolution takes place.