Build fixes for Linux.
Fixing the build for Linux builds with GCC.

The facter header file failed to include facter's collection header so
it was not building (not caught because of cached PCH on a different
platform).

GCC has issues with the implicit casts of long long to value.  Adding
some static casts to make the cast to int64_t explicit, which allows the
implicit cast to value.