Fix lexing multiline comments as regex literals.
A change to the regex literal accidentally omitted a check to see if the regex
should not match a '*' following the opening '/' character, which should be
lexed as a comment.  This resulted in multiline comments being lexed as regular
expression literals.

The fix is to add the check for the '*' character back in.