Fix default code directory for non-root user.
Fixing the default code directory for non-root users, which should be under
`~/.puppetlabs/etc/code`.

The bug was that there was a check to see if the global location existed, and
if it did, to use that location even when non-root.

The fix is to remove that check and only use the global location for root.
This matches the behavior of Puppet.