Make hash stable in insertion order.
Previously hash was implemented solely with an unordered_map, which maintains
no implicit ordering of the elements (hence the name).  However, Puppet's
current behavior is to rely on Ruby's hash type, which maintains insertion
order when enumerating elements of the hash.

This commit changes the implementation to an indexed linked list so that the
hash can be enumerated in insertion order and maintain an O(1) lookup and
erase.