Fix plus operator for Hash + Array.
The current implementation is incorrectly treating a subarray of size 1 as not
being a hash key-value pair (the value would implicitly be undef).

This means the following causes an error instead of inserting an element with
an undef value:

```puppet
{} + [[foo]]
```

Results in an error instead of `{foo => undef}`

This fixes the plus operator such that it respects subarrays of size 1 and
implicitly treats the value as undef.