Implement EPP lexing and parsing.
This implements EPP in the lexer by utilizing Spirit.Lex lexer states.  When
lexing for EPP, the initial state is "EPP", which looks for comment and open
tags.  Certain tags, such as '<%', '<%-', and '<%=', transition the lexer to the
default state, allowing normal lexing of Puppet tokens.  A special "EPP end"
token is emitted to the token stream only for the '<%=' tag, which allows the
parser to know when it should stop parsing of a "render expression".  No tokens
are emitted to the stream for other opening tags and their corresponding end
tags, allowing the parser rules to parse as if the token stream contains only
normal Puppet tokens.

The AST has been amended to contain optional EPP parameters at the root, along
with three new expressions: render string, render expression, and render block.
Render string is intended to render a string directly to the output stream.
Render expression is intended to render the result of a single expression.
Render block is intended to render the result of a block (a list of
statements).

An evaluator that can render an EPP syntax tree will be included in an upcoming
commit; the existing expression evaluator will throw if it encounters a render
expression.