Fix parser rule for resource expression.
The resource expression rule was greedily matching against a type without a
access expression, causing a failure when the parser encountered a resource
defaults expression.

The fix is to change the rule to require the type to have at least one access
expression to match the resource expression rule.  This allows the rule to
gracefully fail to parse and allow the resource defaults expression rule to
match.

Also a minor fix to the debug output of `inline_epp` (missing new line).