Fix reporting of missing '}' for interpolated parsing.
Currently the only use of the "position" parser is for string interpolation.
However, a `get_info` helper has not been written so the error message prints
out the typeinfo for the parser class instead of the expected token.

This adds a specialization of `get_info` such that the expected token is
printed instead of the typeinfo.