Prevent EPP from parsing with a missing EPP end tag.
Currently this is valid EPP to parse:

```
<%
```

This is because EPP open tags `<%` and `<%-` do not output a corresponding end
tag token in the token stream.  To prevent this from parsing, the parser should check
to see if the ending iterator was not in the "EPP" state (i.e. it is currently
processing an EPP tag using the normal lexer state).  If it is in the EPP
state, throw a parse exception letting the user know EOI was encountered before
the closing tag was.