Fix scope not keeping a reference on the AST for variable context.
Previously the scope was storing the AST context where a variable was assigned.
This caused problems with scopes that outlived the ASTs that evaluated
assignment expressions.

This changes what is stored to just keep the shared reference to the file path
and line without having to keep a reference on the AST.