Switch `long double` to `double`
On 64-bit Windows, as well as all OSes on POWER, ARM, and SPARC, `long
double` is actually just a plain old `double`. It's generally just
misleading to anyone reading your code, for very little gain in the
couple cases where it is actually a different type. It's better to use
a type with more consistent behaviors across all platforms.

Even on x86 Linux and OSX (where it's 80-bits), it sucks. It can't be
passed in register arguments, and it actually takes up 16 bytes for
alignment reasons - both in structures and on the stack.

Looking further forward, if we wanted to try to make float evaluation
more consistent across platforms we could start controlling rounding
modes etc. We can't do that if we're using differently-sized floats on
different OSes.