Parser refactoring to support AST range decorations.
This refactoring updates the lexer, AST, and parser to annotate the AST with
needed range information.  This will be needed for AST serialization in the
future, but it also immediately provides for "highlighting" of the error in
error messages (using a "^~~~" squiggle).

The context parser has been replaced with two parsers: begin and end.  The
former parses the beginning position of a token and the latter parses the
ending position of a token.  This allows the rules to populate needed position
information depending on rule specific logic.

This also fixes string interpolation of heredocs not populating the right
context on errors. Previously the start of the string token (the @ for heredocs)
was used as the start of the string's value, but this was incorrect.  The lexer
now properly records the range of the string value itself so that the context
can be calculated from where the string's value starts and not where the token
starts.

Closes GH #98.