(NPUP-1) Fix nested expressions to include parenthesis in context.
Nested expressions were not including the parenthesis in their range, resulting
in highlighting of expressions like `(1  + 2) + 'foo'` to start at the 1 and go
all the way to the ending quote.

This is fixed by adding a new node to the AST called `nested_expression`.  This
node stores the position of the opening and closing parenthesis.  Primary
expressions now contain a `nested_expression` instead of an `expression`.