(NPUP-1) Fix nested query expressions to include parenthesis in context.
Nested query expressions were not including the parenthesis in their range.

This is fixed by adding a new node to the AST called `nested_query_expression`.
This node stores the position of the opening and closing parenthesis.  Primary
query expressions now contain a `nested_query_expression` instead of an
`query_expression`.