(NPUP-1) Fix missing error when there is no matching node.
If a node expression is encountered, it should be an error if a matching node
expression can't be found for the compilation node (no default node or node
that matches the node names).

This was a regression that happened during a recent refactoring.

Also fixes string and regex hostnames not evaluating to the expected strings
when `to_string` was called.