(NPUP-4) Fix type::parse to not take an evaluation context.
Currently `type::parse` takes an evaluation context to work.  This means that
parsing a type specification string uses the current evaluation context,
allowing assignment of variables in the calling scope and catalog modifications,
which should not be allowed.

This commit changes `type::parse` to use a new, empty evaluation context and
changes the parsing to populate a postfix expression rather than a whole AST.

Complex expression evaluation is still allowed, but any operations requiring
access to the node or catalog are not valid in a type specification.