(NPUP-5) Refactor function dispatch to use type system.
This refactors function dispatch to make use of the Puppet type system to
describe the function signatures and validate argument types.  It
also makes the dispatcher dynamic in the sense that new functions can be
registered at runtime; this will be needed to implement functions written in
the Puppet language in the future.

The dispatcher now makes use of function descriptors: a descriptor represents a
single function and has dispatch descriptors that describe an overload's
signature and a callback to invoke if the signature matches.

Descriptors contain logic of finding a matching overload and invoking it.  If
the function call could not be dispatched, a check is done for invalid argument
counts, block parameter counts, or argument types and an exception is thrown. This
makes it so that individual functions don't have to do repetitive validating of
all their arguments.

Note that individual function implementations have transitioned from having to
visit their arguments so that desired overload can be deduced to simply having
dedicated callbacks for a particular overload.

This may be extended for operator dispatch in the future.