(maint) Fix how the Tuple type handles min/max parameters.
The `Tuple` type was not properly handing the min and max parameters.

The min default should be `0`.
The max default should be `numeric_limits<int64_t>::max()`.

Also fix that the evaluator was accepting an empty set of types when evaluating
the access operation.

At least one type is required.