(NPUP-5) Fix incorrect lexing of regex tokens.
Regex lexing and forcing the lexing of a '/' following certain tokens was
incorrect.

The pattern for parsing regex tokens was matching on `/*.../`, which should instead be
treated as an unclosed multiline comment token.  This is fixed by checking for
a leading `/*` and changing the token to unclosed comment.

Additionally, the pattern for checking to see if there exists a `/` following
certain tokens was requiring whitespace following the slash.  This prevented
the correct lexing of expressions such as `6/3/1`, which was lexed as a number,
followed by a regex, followed by a number (instead of number, division, number, division,
number).  This was fixed to only checking the current line and require anything
but a '*' character following the '/' to match.

Also added a missing lexer test for regex tokens.