(NPUP-5) Fix error logging functions to not be treated as failures.
Currently the Puppet logging functions that log at error level or higher cause the
compiler to return a non-zero exit status and report failure despite not
actually failing the compilation.

This commit fixes the logger so that errors can be logged from the Puppet logging
functions without causing the compiler to treat them as a failure.

This commit also removes the recursion used in the logger as it was
unnecessary; it has been replaced with the "initializer_list parameter unpack"
trick.