(NPUP-14) Evaluate parameter default values in new match scope.
When evaluating parameter default value expressions, a new match scope should
be used to prevent match scope variables "leaking" into the evaluation of the
following parameter defaults or into the body of the class, defined type, or
function.

This commit extracts out where the default value was evaluated and wraps it
with a utility member function that creates a new match scope.  Creating a new
match scope is efficient, so this will add no perceivable overhead to the call
evaluator.

This also fixes the same issue when evaluating the type expressions for
parameters.