(NPUP-15) Implement operator dispatch in the dispatcher.
Currently the evaluator handles dispatching binary and unary operator
"calls" and each operator's implementation has to handle validating the
values given to them.  Now that function call dispatching uses the Puppet type
system to do the dispatch, this commit lays the groundwork to unify operator
call dispatch with function call dispatch.

This commit modifies the dispatcher to support dispatching operator calls.

This commit does not include porting the existing operator implementations over
to using the dispatcher as each will be individually ported.

Once all operators have been ported, the evaluator code will be cleaned up so
that it relies completely on the dispatcher.