(NPUP-15) Fix incorrect usage of `numeric_limits<double>::min`.
This corrects the few places where `numeric_limits<double>::min` was being used
to expect the "lowest" value instead of the minimal finite value.  This is a
confusing part of the stdlib where `min` means something different to floating
point values compared to integral types.

Fixed by replacing the `min` calls with `lowest`.

The `max` function is still used because there is no `highest`.