(NPUP-18) Implement string interpolation in the lexer.
Previously, the lexer would naively match anything between double quotes and
interpolated heredoc tags as part of a single string token.  It was then up to
the evaluator to parse the string and perform string interpolations.  However,
this is not what the Puppet language supports in terms of interpolation.
Puppet performs interpolation in the lexer and parser, resulting in AST nodes
containing children comprising the string (literal text, variables, and
expressions to render).

This commit completely changes how string interpolation is performed by the
native compiler.

Firstly, string tokens are emitted only for single quoted strings and
non-interpolated heredocs.  Double quoted strings and interpolated heredocs
result in a "string start" token, followed by any number of literal text
tokens, variable tokens, or an "interpolation sequence".  An interpolation
sequence has a interpolation token, any number of normal lexer tokens, and an
interpolation end token.  Interpolated strings without any variables or
interpolation sequences are transformed to non-interpolated string expressions
by the parser.

The lexer is also now responsible for performing lookahead for interpolation
sequences where the first immediate token could be a variable (e.g. `"${foo}"`).

Additionally, the lexer now performs transformations of escape characters,
which was previously handled by the, now removed, "string interpolator".

Variable name validation is now done in the lexer.

Note that neither the lexer nor the parser aligns heredoc text containing a
margin specifier; that is still handled by the evaluator.  To fix this, we'll
need a semantic action when matching a string / interpolated string rule and
align the text in the resulting AST object.  It is done this way because the
lexer outputs a stream of tokens; the margin isn't known until the "string end"
token is encountered for the heredoc.

Secondly, the parser rules have been updated to produce an "interpolated
string" expression when interpolated string tokens are encountered.

Lastly, the evaluator has been updated to render the interpolated string
(similar to how it renders EPP expressions).  This removes the now-defunct
"string interpolator" that was previously responsible for parsing the string
and replacing interpolation expressions.

Also fixed in this commit:

* Lexer exceptions now contain the range instead of just the starting iterator;
  this allows for better error message highlighting.
* Resource titles should support any expression that results in a string.