(NPUP-17) Extract 'make_absolute' out of compile command.
This extracts `make_absolute` out of the compile command so it can be used in
other places.

This also fixes the function to use Boost's `lexically_normal` instead of
`weakly_canonical` as the latter will resolve symbolic links and the intention
was to normalize the path (i.e. remove . and .. entries in the path).