(NPUP-17) Change 'environment-dir' option to 'environment-path'.
The old 'environment-dir' option was a multi-option, allowing users to specify
the option multiple times to build a list of directories to search.

However, this is inconsistent with how Puppet defines the environment path
setting.

This commit changes the option to a single 'environment-path' option which
accepts a separator-delimited list of paths to search for environments.

A future commit will also address the incorrect 'module-path' option.