(NPUP-17) Change 'module-dir' option to 'module-path'.
The old 'module-dir' option was a multi-option, allowing users to specify
the option multiple times to build a list of directories to search.

However, this is inconsistent with how Puppet defines the base module path
setting.

This commit changes the option to a single 'module-path' option which
accepts a separator-delimited list of paths to search for modules.