(NPUP-17) Fix compile command exit status.
Fixing the compile command exit status to return non-zero on an exception
rather than keeping track of failure in the logger.  The problem with doing it
the previous way was that the logging level setting affected whether or not the
command returned EXIT_SUCCESS or EXIT_FAILURE.

The command should signal success provided that there were no exceptions
encountered during command execution.