(NPUP-17) Change --log-level option to a string.
Previously the `--log-level` option was a logging::level and required an
`operator>>` to extract the option from a string.  The problem with this is
that the parse command requires a subset of the supported options.  For it to
be able to override what the supported log levels it must be able to read the
option as a string.

This removes the now unnecessary `operator>>` in favor of just handling to
conversion from string to log level in `command::get_level`.