(maint) Fix failure to find existing resource for error message.
An attempt to find an existing resource after failing to declare a resource was
due to reusing a string that had been previously moved.  As a result, a less
helpful "expected previous resource" error message would be displayed.

This fixes the bug by not moving the type into the call to `catalog.add` so it
can be used in the error case.