(NPUP-26) Fix class evaluation when parent class starts with '::'.
When a parent class starts with '::', the lookup of the parent's scope is not
done with a normalized name (without the '::').  This results in a null scope,
which causes the inheriting class' scope to throw an exception.

This commit fixes the problem by using the title from the returned resource
when declaring the parent class; the title has been correctly normalized.