(NPUP-27) Fix detection of circular class inheritance.
When a class inherits from another that, in turn, inherits from the original
class, forming a circular reference, the evaluation would fail with a confusing
message about not being able to locate the parent scope.

This commit improves that error message so that users know they have a circular
inheritance.