(maint) Fix failure to stop at top scope for override check.
For evaluating a resource override, a check is done to see if the resource
being overridden is contained in a parent scope.  However, the existing check
fails to stop before checking the top scope; this results in allowing any class
to modify the existing attributes of resources declared in the top scope.

The fix is to simply stop when we reach a scope without a parent.