(NPUP-6) Implement private functions.
Implementing private functions by supporting the 'private' keyword before
function definitions.

Private functions can only be defined in modules; an attempt to use
the private keyword in functions defined in a main manifest will result in
an error.

The function's descriptor enforces that the caller comes from the same module;
this is done by comparing the function definition's module to the that of the
call expression (ASTs are associated with their modules when manifests are
imported into an environment). This happens when the call is dispatched and
does not rely on a separate loader to function.