(maint) Update CI docker image to use Alpine instead of Arch.
This updates the docker image used in CI to be based off of Alpine instead of
Arch.  This reduces the size of the image needed to be downloaded upon each CI
build.

This also changes the CI image to be 'puppetcpp-ci' so that the 'puppetcpp' can
be an image with a pre-built compiler to experiment with.