(maint) Add value name to hash value.
For runtime values, adding the name of the value to the computed hash value
to ensure undef and default don't share the same computed value (which
previously was just 0).

The name of the value is computed only once and then stored for merging in
with other computed values depending on what the value represents.