(NPUP-34) Implement type alias evaluation.
This implements type alias evaluation and resolution.

Type aliases are resolved when an "instance" of the type is created.  The
resolution is performed once and the result cached and reused on subsequent
references to the alias.

The type system had to be modified to allow for recursive resolutions when
printing types and determining if an alias contains a cycle.