(NPUP-39) Implement is_assignable for all types in the type system.
Previously a `is_specialization` function was implemented to implement the
comparison operators on types under the assumption the operators were answering
the question of "is this type more or less specific?".  However, the question
the operators answer is actually "is this type assignable?".  As such, this
commit replaces `is_specialization` with `is_assignable` and adheres to the
logic of `assignable?` from the Ruby implementation.

Additionally, a proper "recursion guard" was implemented to prevent infinite
recursion on type aliases which was missing for instance checking.

Since all types except `Variant` and type aliases returned `true` for
`is_real`, the `is_real` implementations were removed in favor of a much
simpler member function on `values::type` that does the same check.

This commit comes with a comprehensive test that checks the comparison
operators on every type in the type system (except for `Iterator` since one can
not yet be instantiated without a function that returns an `Iterator`
instance).

Minor bugs encountered and also fixed:

* A few types with `from` parameters were defaulting to
  `numeric_limits<T>::min()` instead of 0.
* In the access operator, a few errors were incorrectly referencing an AST
  context that was out of range, causing a crash.